fix(build): dependency issues with new awaitility
Awaitility version was bumped due to it being declared in the Camel BOM,
this version is using an older version of hamcrest which we now exclude
from the build not to interfere with the newer version at different
Maven coordinates.

Also reverts back to using the JUnit 4 dependency in the project
generator since using Jupiter version brings about the duplication of
classes.