fix(camel-k): force camel-api version
For some undetermined reason, version of Camel in the Camel K plugin is
set to 3.1.0-SNAPSHOT and the version of camel-api is set to 3.0.0-RC3.

With the refactoring in CAMEL-14498[1] we need to make sure that the
`camel-api` version is aligned with the version of Camel that's being
used by the plugin.

[1] https://issues.apache.org/jira/browse/CAMEL-14498