chore(c/http): migrate HTTP connector to Camel 3.1
Since in Camel 3.0 `http4` and `https4` components were renamed to
`http` and `https` this makes the symmetrical change in Syndesis HTTP
connector.

Migration was added to make sure that existing integrations containing
the HTTP connector would be valid with this change. Since it might take
a while for the Camel 3 branch to stabilize higher schema version was
chosen to prevent conflicts when rebasing.